library to calculate host-guest int energy

+++ src/ljsampler.py
@@ -1,4 +1,5 @@
 import os 
+import errno
 from pymatgen.io.cif import CifParser
 from pymatgen.core import PeriodicSite 
 import pandas as pd
@@ -22,6 +23,8 @@ class load_coordinates():
         self.df_FF = pd.DataFrame(FF_content, columns=columns).astype({'epsilon':float, "sigma":float})
         if self.df_FF.loc[0,'interaction form'] != 'lennard-jones':
             raise ValueError("Unsupported interation form: only Lennard-Jones interactions are supported")
+        if not os.path.exists("Coordinates"):
+            raise FileNotFoundError(errno.ENOENT, os.strerror(errno.ENOENT), "Coordinates")
         self.atoms = atoms
         self.cutoff = cutoff
         self.temperature = temperature
@@ -29,40 +32,57 @@ class load_coordinates():
         if not os.path.exists("Energies"):
             os.mkdir("Energies")
 
-    def sample(self, structure_name, unitcell, lattice_matrix, min_distance=2.0):
+    def evaluate(self, structure_name, supercell, lattice_matrix, min_distance=2.0, precision=2):
         """Calculates the average energies/ minimum energy/ boltzmann average energies/ list of energies
 
         Args:
             structure_name (str): name of the structure as referenced in the cif file in the Raspa directory
-            unitcell (str): supercell unitcell to prevent self interaction at 12 angstrom cut off
+            supercell (list): list of ints representing the size of each unicell vector for the supercell
             lattice_matrix (numpy.2Darray): 3x3 array representing the unitcell vectors in the cartesian coordinate 
             system of Raspa2 and Zeo++ (different from the one used in pymatgen)
+        
+        Returns:
+
 
         """
 
-        df_voro = pd.read_csv(os.path.join("Coordinates",structure_name+'.csv'), sep=" ")
-        df_voro = df_voro[df_voro["dist_to_nearest"]>min_distance]
-        cif_file = open(os.path.join(self.RASPA_DIR,'share/raspa/structures/cif',structure_name+".cif"),"r").read()
-        cif_file = cif_file.replace('_atom_site_type_symbol','_atom_site_label')
-        parser = CifParser.from_string(cif_file)
-        structure = parser.get_structures()[0]
-        structure.make_supercell([int(u) for u in unitcell.split()])
-        df_voro["cartesian_coordinates"] = df_voro.apply(lambda row: np.array([float(row.x),float(row.y),float(row.z)]), axis=1)
-        inv_lattice_matrix = np.linalg.inv(lattice_matrix)
-        df_voro["fractional_coordinates"] = df_voro["cartesian_coordinates"].apply(lambda coord: np.matmul(inv_lattice_matrix, np.array(coord)))
+        coord_path = os.path.join("Coordinates",structure_name+'.csv')
+        if os.path.exists(coord_path):
+            df_voro = pd.read_csv(coord_path, sep=" ")
+            df_voro = df_voro[df_voro["dist_to_nearest"]>min_distance]
+            cif_file = open(os.path.join(self.RASPA_DIR,'share/raspa/structures/cif',structure_name+".cif"),"r").read()
 
-        accessible_mean_energy, min_energy, boltz_energy = [], [], []
-        for molecule in self.atoms:
-            df_voro["pymat_site_%s"%molecule] = df_voro["fractional_coordinates"].apply(lambda frac_coord: PeriodicSite({molecule:1}, frac_coord, structure.lattice))
-            df_voro["Energy_%s"%molecule] = df_voro["pymat_site_%s"%molecule].apply(lambda site: self.lennard_jones(site, structure, self.cutoff))
-            accessible_mean_energy.append(df_voro[df_voro["Energy_%s"%molecule]<0]["Energy_%s"%molecule].mean())
-            min_energy.append(df_voro["Energy_%s"%molecule].min())
-            df_voro["exp_energy_%s"%molecule] = np.exp(-df_voro["Energy_%s"%molecule]/(self.R*self.temperature))
-            Sum_exp = df_voro["exp_energy_%s"%molecule].sum()
-            df_voro["pond_energy_%s"%molecule] = df_voro["exp_energy_%s"%molecule] * df_voro["Energy_%s"%molecule] / Sum_exp
-            boltz_energy.append(df_voro["pond_energy_%s"%molecule].sum())
-        df_voro.to_csv('Energies/%s.csv'%structure_name)
-        return accessible_mean_energy, min_energy, boltz_energy
+            if '_atom_site_label' in cif_file:
+                if '_atom_site_type_symbol' in cif_file:
+                    cif_file = cif_file.replace('_atom_site_label','_atom_site_lab')
+                    cif_file = cif_file.replace('_atom_site_type_symbol','_atom_site_label')
+            else:
+                if '_atom_site_type_symbol' in cif_file:
+                    cif_file = cif_file.replace('_atom_site_type_symbol','_atom_site_label')
+                else:
+                    raise KeyError("_atom_site_type_symbol and _atom_site_label not in the cif file, please check the format")
+
+            parser = CifParser.from_string(cif_file)
+            structure = parser.get_structures()[0]
+            structure.make_supercell(supercell)
+            df_voro["cartesian_coordinates"] = df_voro.apply(lambda row: np.array([float(row.x),float(row.y),float(row.z)]), axis=1)
+            inv_lattice_matrix = np.linalg.inv(lattice_matrix)
+            df_voro["fractional_coordinates"] = df_voro["cartesian_coordinates"].apply(lambda coord: np.matmul(inv_lattice_matrix, np.array(coord)))
+
+            accessible_mean_energy, min_energy, boltz_energy = [], [], []
+            for molecule in self.atoms:
+                df_voro["pymat_site_%s"%molecule] = df_voro["fractional_coordinates"].apply(lambda frac_coord: PeriodicSite({molecule:1}, frac_coord, structure.lattice))
+                df_voro["Energy_%s"%molecule] = df_voro["pymat_site_%s"%molecule].apply(lambda site: self.lennard_jones(site, structure, self.cutoff))
+                accessible_mean_energy.append( round(df_voro[df_voro["Energy_%s"%molecule]<0]["Energy_%s"%molecule].mean(), precision) )
+                min_energy.append( round(df_voro["Energy_%s"%molecule].min(), precision) )
+                df_voro["exp_energy_%s"%molecule] = np.exp(-df_voro["Energy_%s"%molecule]/(self.R*self.temperature))
+                Sum_exp = df_voro["exp_energy_%s"%molecule].sum()
+                df_voro["pond_energy_%s"%molecule] = df_voro["exp_energy_%s"%molecule] * df_voro["Energy_%s"%molecule] / Sum_exp
+                boltz_energy.append( round(df_voro["pond_energy_%s"%molecule].sum(), precision) )
+            df_voro.to_csv('Energies/%s.csv'%structure_name)
+            return accessible_mean_energy, min_energy, boltz_energy
+        else:
+            return np.nan,np.nan,np.nan
 
 
     def lennard_jones(self, atom_g, structure_h, cutoff, shifted=True):
@@ -82,7 +102,8 @@ class load_coordinates():
         E = 0.0
         shift = 0.0
         row_g = self.df_FF[self.df_FF['atom type'] == atom_g.specie.symbol]
-        for atom_h in structure_h:
+        neighbors = structure_h.get_neighbors(atom_g, cutoff)
+        for atom_h in neighbors:
             distance = atom_h.distance(atom_g)
             if  distance < cutoff:
                 row_h = self.df_FF[self.df_FF['atom type'].str.strip('_') == atom_h.specie.symbol]
